[#53]refactor: name 클래스에 주석 추가, 폴더 이름 변경
